requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose==3.3.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
pydantic-settings>=2.0.3
python-multipart==0.0.6
aiohttp==3.9.5
itsdangerous==2.1.2
passlib==1.7.4
bcrypt==4.0.1